requirements.txt
requests
simplejson
uuid
flask
flask-cors
python-dateutil
